# public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Autenticação",
        "description": "Documentação da API de autenticação usando Swagger no Lumen",
        "version": "1.0.0"
    },
    "paths": {
        "/login": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Realiza o login do usuário",
                "operationId": "2dd578ff7aba2721293dac66833b27d1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "usuario@email.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "senha123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login realizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Login realizado com sucesso"
                                        },
                                        "data": {
                                            "properties": {
                                                "token": {
                                                    "type": "string",
                                                    "example": "eyJhbGciOiJIUzI1..."
                                                },
                                                "user": {
                                                    "properties": {
                                                        "uuid": {
                                                            "type": "string",
                                                            "example": "123e4567-e89b-12d3-a456-426614174000"
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "João Silva"
                                                        },
                                                        "email": {
                                                            "type": "string",
                                                            "example": "usuario@email.com"
                                                        },
                                                        "avatar": {
                                                            "type": "string",
                                                            "example": "https://exemplo.com/avatar.jpg"
                                                        },
                                                        "phone": {
                                                            "type": "string",
                                                            "example": "+55 11 99999-9999"
                                                        },
                                                        "status": {
                                                            "type": "string",
                                                            "example": "ativo"
                                                        },
                                                        "is_admin": {
                                                            "type": "boolean",
                                                            "example": false
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciais inválidas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Credenciais inválidas"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/categories": {
            "get": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Listar todas as categorias",
                "operationId": "ead5fb1cf2a7453acb2e37928a1b1eaf",
                "responses": {
                    "200": {
                        "description": "Lista de categorias"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Criar nova categoria",
                "operationId": "bb02f0320882db37f99574196577e446",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Categoria criada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/categories/{uuid}": {
            "get": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Buscar categoria por UUID",
                "operationId": "b9617e7afa6b3fafc37cbf883aa2b24f",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "UUID da categoria",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Categoria encontrada"
                    },
                    "404": {
                        "description": "Categoria não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Atualizar categoria",
                "operationId": "d4818e735365c518865a01c68bd7a6dc",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "UUID da categoria",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Categoria atualizada"
                    },
                    "404": {
                        "description": "Categoria não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Excluir categoria",
                "operationId": "a901381c34a7704c317ebe68b1c5ab6b",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "UUID da categoria",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Categoria excluída com sucesso"
                    },
                    "404": {
                        "description": "Categoria não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/stock/metrics": {
            "get": {
                "tags": [
                    "Estoque"
                ],
                "summary": "Retorna métricas do estoque",
                "operationId": "64787a297fb74d3ef1274af4be49ca54",
                "responses": {
                    "200": {
                        "description": "Métricas retornadas com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "purchases_total": {
                                            "type": "integer",
                                            "example": 150
                                        },
                                        "categories_total": {
                                            "type": "integer",
                                            "example": 12
                                        },
                                        "provider_total": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "employees_total": {
                                            "type": "integer",
                                            "example": 8
                                        },
                                        "monthly_purchases": {
                                            "type": "object",
                                            "example": {
                                                "Nov": 1234.56,
                                                "Oct": 980.25,
                                                "Sep": 1100
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "tags": [
        {
            "name": "Autenticação",
            "description": "Endpoints relacionados à autenticação do usuário"
        },
        {
            "name": "Categorias",
            "description": "Categorias"
        },
        {
            "name": "Estoque",
            "description": "Estoque"
        }
    ]
}